Feature：buffer限制最大容量；file写完一块，sync一块

+++ src/main/java/com/fineio/v3/connector/Connector.java
@@ -36,6 +36,6 @@ public interface Connector {
     /**
      *
      */
-    byte getBlockOffset();
+    int getBlockOffset();
 
 }
